Remove the app/ directory
Now it's time to get interesting. Like every folder generated by `rails new`, the app directory is only a suggestion - we can delete it entirely, and move the parts we need directly into `application.rb`. Essentially, we're just "inlining" several files into one.

This app isn't going to be using any assets or ActionCable, so we can obviously delete those files. Layouts are optional, too!

Rails 5 introduces the concept of an ApplicationRecord - like ApplicationController, it's simply a place for defining behavior that applies to *all* models in your application. It's more for plugin authors to hook into, and it's not strictly necessary (unless we were using a plugin that expected us to have an ApplicationRecord, of course).

198 LOC / 36 files