Only load ActionController. Remove initializers, environments.
OK, this is a big one!

The most important change is the removal of `require rails/all`. If we look at `rails/railties/lib/rails/all.rb`, we'll see that it's a very simple file that just loads all of the component parts of the Rails framework. Each of these parts is actually highly modular, and many (even most) Rails applications probably don't use all parts. We only need one of Rails' components for a Hello World app - `action_controller`. Here's what we *won't* be loading:

* ActiveRecord. We don't need a database.
* ActionMailer. We're not sending emails.
* ActiveJob. No background jobs for us!
* ActionCable. WebSockets? More like LameSockets.
* Rails Test/Unit. We're not testing
* Sprockets. No assets!

Note, though, that many of these components *are* necessary for any non-trivial webapp. When was the last time you were paid to work on an app that doesn't use a database? However, Sinatra (as an example) provides *none* of these facilities!
Since we're no longer loading ActiveRecord or ActionMailer, we need to remove any config settings that refer to those components. While we're at it, we can just remove the development and test environments. Environment-specific config files, while very useful, aren't strictly necessary for a Rails application.

We've also deleted all of the initializers - if we needed any of these files, we could just inline them into config/application.rb.

76 LOC / 17 files